Remove AWS from the naem of audit log dynamo gateway
