+++ scripts/utils/check-records-status.ts
@@ -11,7 +11,7 @@
  */
 
 import fs from "fs"
-import { AwsAuditLogDynamoGateway } from "../../src/audit-log-api/src/gateways/dynamo"
+import { AuditLogDynamoGateway } from "../../src/audit-log-api/src/gateways/dynamo"
 import CalculateMessageStatusUseCase from "../../src/audit-log-api/src/gateways/dynamo/AuditLogDynamoGateway/CalculateMessageStatusUseCase"
 import { isError } from "../../src/shared-types/src"
 import { Lambda } from "../../src/shared/node_modules/aws-sdk"
@@ -58,7 +58,7 @@ async function setup() {
   }
 }
 
-const dynamo = new AwsAuditLogDynamoGateway(dynamoConfig, dynamoConfig.TABLE_NAME)
+const dynamo = new AuditLogDynamoGateway(dynamoConfig, dynamoConfig.TABLE_NAME)
 
 async function run() {
   let totalProcessing = 0
