+++ src/audit-log-api/src/handlers/getEvents.ts
@@ -4,14 +4,14 @@ import type { AuditLogEvent, PromiseResult } from "shared-types"
 import { isError } from "shared-types"
 import createAuditLogDynamoDbConfig from "../createAuditLogDynamoDbConfig"
 import createAuditLogLookupDynamoDbConfig from "../createAuditLogLookupDynamoDbConfig"
-import { AwsAuditLogDynamoGateway, AwsAuditLogLookupDynamoGateway } from "../gateways/dynamo"
+import { AuditLogDynamoGateway, AwsAuditLogLookupDynamoGateway } from "../gateways/dynamo"
 import { FetchEventsUseCase, parseGetEventsRequest } from "../use-cases"
 import LookupEventValuesUseCase from "../use-cases/LookupEventValuesUseCase"
 import { createJsonApiResult } from "../utils"
 
 const auditLogConfig = createAuditLogDynamoDbConfig()
 const auditLogLookupConfig = createAuditLogLookupDynamoDbConfig()
-const auditLogGateway = new AwsAuditLogDynamoGateway(auditLogConfig, auditLogConfig.TABLE_NAME)
+const auditLogGateway = new AuditLogDynamoGateway(auditLogConfig, auditLogConfig.TABLE_NAME)
 const auditLogLookupGateway = new AwsAuditLogLookupDynamoGateway(auditLogLookupConfig, auditLogLookupConfig.TABLE_NAME)
 const lookupEventValuesUseCase = new LookupEventValuesUseCase(auditLogLookupGateway)
 const fetchEvents = new FetchEventsUseCase(auditLogGateway, lookupEventValuesUseCase)
