+++ src/audit-log-api/src/use-cases/CreateAuditLogEventUseCase.ts
@@ -1,6 +1,6 @@
 import type { AuditLogEvent, CreateAuditLogEventsResult } from "shared-types"
 import { isError } from "shared-types"
-import type { AuditLogDynamoGateway } from "../gateways/dynamo"
+import type { AuditLogDynamoGatewayInterface } from "../gateways/dynamo"
 import { isConditionalExpressionViolationError } from "../gateways/dynamo"
 import type StoreValuesInLookupTableUseCase from "./StoreValuesInLookupTableUseCase"
 
@@ -35,7 +35,7 @@ const isDuplicateEvent = (event: AuditLogEvent, existingEvents: AuditLogEvent[])
 
 export default class CreateAuditLogEventUseCase {
   constructor(
-    private readonly auditLogGateway: AuditLogDynamoGateway,
+    private readonly auditLogGateway: AuditLogDynamoGatewayInterface,
     private readonly storeValuesInLookupTableUseCase: StoreValuesInLookupTableUseCase
   ) {}
