+++ src/audit-log-api/src/use-cases/CreateAuditLogEventsUseCase.ts
@@ -1,6 +1,6 @@
 import type { AuditLogEvent, CreateAuditLogEventsResult } from "shared-types"
 import { isError } from "shared-types"
-import type { AuditLogDynamoGateway, DynamoUpdate } from "../gateways/dynamo"
+import type { AuditLogDynamoGatewayInterface, DynamoUpdate } from "../gateways/dynamo"
 import { isConditionalExpressionViolationError, isTooManyEventsError } from "../gateways/dynamo"
 import type StoreValuesInLookupTableUseCase from "./StoreValuesInLookupTableUseCase"
 
@@ -35,7 +35,7 @@ const isDuplicateEvent = (event: AuditLogEvent, existingEvents: AuditLogEvent[])
 
 export default class CreateAuditLogEventsUseCase {
   constructor(
-    private readonly auditLogGateway: AuditLogDynamoGateway,
+    private readonly auditLogGateway: AuditLogDynamoGatewayInterface,
     private readonly storeValuesInLookupTableUseCase: StoreValuesInLookupTableUseCase
   ) {}
