+++ src/audit-log-api/src/use-cases/CreateAuditLogsUseCase.ts
@@ -1,6 +1,6 @@
 import type { AuditLog, TransactionFailureReason } from "shared-types"
 import { isError } from "shared-types"
-import type { AuditLogDynamoGateway } from "../gateways/dynamo"
+import type { AuditLogDynamoGatewayInterface } from "../gateways/dynamo"
 import { isConditionalExpressionViolationError, isTransactionFailedError } from "../gateways/dynamo"
 
 interface CreateAuditLogsResult {
@@ -10,7 +10,7 @@ interface CreateAuditLogsResult {
 }
 
 export default class CreateAuditLogsUseCase {
-  constructor(private readonly auditLogGateway: AuditLogDynamoGateway) {}
+  constructor(private readonly auditLogGateway: AuditLogDynamoGatewayInterface) {}
 
   async create(auditLogs: AuditLog[]): Promise<CreateAuditLogsResult> {
     const result = await this.auditLogGateway.createMany(auditLogs)
