+++ src/audit-log-api/src/use-cases/FetchAll.ts
@@ -1,6 +1,6 @@
 import type { AuditLog, PromiseResult } from "shared-types"
 import { isError } from "shared-types"
-import type { AuditLogDynamoGateway } from "../gateways/dynamo"
+import type { AuditLogDynamoGatewayInterface } from "../gateways/dynamo"
 import getMessageById from "./getMessageById"
 import type MessageFetcher from "./MessageFetcher"
 
@@ -11,7 +11,10 @@ type FetchAllOptions = {
 }
 
 export default class FetchAll implements MessageFetcher {
-  constructor(private readonly gateway: AuditLogDynamoGateway, private readonly options: FetchAllOptions = {}) {}
+  constructor(
+    private readonly gateway: AuditLogDynamoGatewayInterface,
+    private readonly options: FetchAllOptions = {}
+  ) {}
 
   async fetch(): PromiseResult<AuditLog[]> {
     const lastMessage = await getMessageById(this.gateway, this.options.lastMessageId)
