+++ src/audit-log-api/src/use-cases/FetchById.ts
@@ -1,10 +1,10 @@
 import type { AuditLog, PromiseResult } from "shared-types"
-import type { AuditLogDynamoGateway, ProjectionOptions } from "../gateways/dynamo"
+import type { AuditLogDynamoGatewayInterface, ProjectionOptions } from "../gateways/dynamo"
 import type MessageFetcher from "./MessageFetcher"
 
 export default class FetchById implements MessageFetcher {
   constructor(
-    private readonly gateway: AuditLogDynamoGateway,
+    private readonly gateway: AuditLogDynamoGatewayInterface,
     private messageId: string,
     private readonly options?: ProjectionOptions
   ) {}
